There's only a single type: Value
Both Relative and Absolute values express the idea of "a calendar value of a contiguous range of units". The only difference is that in an Absolute value, the largest unit is ALWAYS the "Era".

Based on this realization, we can have a *single* type (Value<S, L>) that is used for BOTH relative and absolute values, but absolute values always have L == Era