Retrieve a sub-unit with a particular user-visible number
There is a calendrical difference between "The second hour" and "The hour called '2'". The latter is usually what people want when dealing with calendars, as it maps most closely to what we'd see by looking at a clock.

The "third" hour of a day is from 2AM – 3AM (because the "first" hour is from midnight – 1AM), but the "hour called 3" is from 3AM – 4AM.

Similarly there are times when larger units get skipped, such as leap days or leap months, and these also throw off ordinalities vs cardinalities.