Handle autoupdating region values
When a DateFormatter is given an autoupdating Locale value, the values from that locale will override any custom calendar set on the date formatter, preferring instead the calendar from the autoupdating locale.

This change makes it so that DateFormatters are _always_ built using non-autoupdating versions of the region components. However, the DateFormatterCache will remember which formatters were created from autoupdating Regions, and then un-cache them if something about the autoupdating Region changes.