Allow "exactDate" method to supply missing era
Eras are not always relevant, so it's common to see them omitted when creating dates. However, the "exactMatch" method on Calendar was expecting *all* components to be present, including the era. This method has been updated to supply the missing era, but only if the calendar itself knows that eras are not really relevant.

By default, only the Japanese calendar considers eras to be relevant. Fixes #27