Rename "Unsafe" -> "Strict"
Fixes #20.

There was objection on the forums to the usage of "unsafe", since Swift historically has used that to mean "memory unsafe". After much discussion, I'm replacing it with "Strict", which brings it in line with other usages, and also matches `Calendar.MatchingPolicy`'s `.strict` case.

This change *leaves in* the old initializer that has "unsafe" in it, but marks it as deprecated with a renaming hint and a description of what changed.