TimeDifferences restrict units but do not require them
A TimeDifference is allowed to have missing units (ex: "a year and a day" has no "month" value). Therefore, it should not be using the `requireAndRestrict()` method to guarantee that all intermediate units are present. Therefore, this introduces a new "restrict()" method to remove all disallowed units from a DateComponents, but does not throw an error if allowed units are missing.

This also exposed the issue that `value(for:)` was returning `NSNotFound` for units that are not present in a DateComponents, as opposed to `nil`. Therefore, we specifically filter out `NSNotFound` values from the restricted values. However, this revealed that not all DateComponents have the "required" set of units (specifically, many absolute TimePeriods were missing the .era unit, because their calendar did not require it. Therefore, this *also* introduces the notion of "lenient" units: in the process of requiring a certain set of Calendar.Components, lenient ones may be missing without causing an error to be thrown. So far, the only lenient unit used it the `.era` unit, and only then when the calendar itself says that the era value may be omitted (ie, `isEraRelevent == false`)

All unit tests pass.