Change Fixed<U>'s internal storage
Swift.Duration is a better type to use for date values than Foundation.Date: it has higher precision and lives in a lower framework. This change moves Fixed<U> off of Foundation.Date and onto Swift.Duration, and then constructs Foundation.Date values when needed.

Serialization unit tests still pass, as the deserialization of Fixed<U> is adaptive to the incoming format.