First pass at integrating weeks
This requires some adjustment to the unit protocols. Weeks are shorter than months, but longer than days. This means we need a set of protocols that express a "strictly greater than" or "strictly less than" relationship, so we have `LTMonth` and `GTDay` now (for example).

Additionally, a lot of the current methods assume that boundaries between units line up, which is not the case for weeks. Units with aligned boundaries (like how the start of a month is also the start of a day or and hour) are called "standard units". Units that are layered on top of standard units (weeks and quarters) are "pseudo units". Methods and properties that rely on standard alignment now require constraints to `StandardUnit`, so that they can be limited to only standard values. This allows for selective exposure of APIs on `Fixed<Week>`.

Also, Weeks express a calendar component that is not a standard component. Therefore, the `Unit` protocol needs two values: the unit it actually represents, and the unit it needs to be present in order to extract meaningful information. For standard units, those are the same. For pseudo-units, they're different. `Week` represents the `.weekOfYear`, but requires `.day` to be present.